requirements.txt
#pip
Django>=1.7,<1.8
django-haystack>=2.3,<2.4
whoosh>=2.5,<2.6
mistune==0.4.1
Pillow>=2.6,<2.7
django-infinite-scroll-pagination>=0.1.3,<0.2
django-djconfig>=0.1.7,<0.2
django-debug-toolbar
pytz
South>=0.7.4
djangorestframework
markdown
django-filter
django-role-permissions
tox==1.7.0
model-mommy==1.2.1
django-discover-runner==1
pysqlite==2.6.3

#easy_install
sudo easy_install django-rosetta